+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Added `Delete Chat Messages` endpoint
 * Added `Get User Chat Color` and `Update User Chat Colors` endpoint
 * Added `Add Channel Moderator` endpoint
+* Added `Get VIPs`, `Add Channel VIP` and `Remove Channel VIP` endpoint
 
 ## [v0.6.1] - 2022-04-29